feat(gh): allow latest arg to be an enum
The possible values for `latest` are a trinary: "true", "false", and a
magic auto value that happens when the arg isn't specified:

> --latest
>        Mark this release as "Latest" (default [automatic based on date and version]).
>        --latest=false to explicitly NOT set as latest

Today the value is only a boolean, and so there's no way at the moment
to force latest to be false. By updating the arg type to a trinary enum,
then we can express this constraint properly.

Signed-off-by: Justin Chadwell <me@jedevc.com>